Fixes PR 2542384 SnapshotManager.CreateVolumeFromSnapshot created a download record with a name of format download-<snapshot ID>-<uuid> however, when it went to check the status of the record it searched for download-<snapshot ID> and did not find it.  The error from Get() was being ignored.  The record returned was empty and the code checks for phase == completed or failed and continues to wait if the phase is neither of those.
Changed to generate 1 string for the name of the record and use it in both places.  Added error checking and logging to
monitor polling of the record.

Additionally, if the PollImmediateInfinite returned an error the error was ignored and we attempted to create a new ID
from download.Status.VolumeID.  Since this was an empty string, the create failed but the error message we would see in the logs
was incorrect.  Fixed to check for the error from PollImmediateInfinite.

Signed-off-by: Dave Smith-Uchida <dsmithuchida@vmware.com>