Fix to get vSphere credentials for TKG cluster.
The velero vsphere plugin piggy backs on the csi secret in the worker nodes to
get VC credentials. For a Vanilla Kubernetes cluster, the secret is stored in
vsphere-config-secret in the kube-system namespace. However, for the TKG cluster,
the secret name is not the same and so the call to retrive the VCConfigSecret
fails.
Made the following change:
1. Added a new secret location csi-vsphere-config, that is valid for TKG
cluster, to check for csi-vsphere.conf with the VC credentials.

2. Updated the VC config params for the following checks in csi-vsphere.conf:
   - added a default port if the port is missing in the CSI config file. This is
     similar to CSI code behavior.
   - Updated the config values correctly even if "" are missing in the config file.

3. Create a new IDV preotected entity by passing all the secret params so astrolabe
   code does not have to retrieve them again.

Testing Done:
  Manually tested with TKG cluster and vanilla cluster on vsphere.
  https://container-dp.svc.eng.vmware.com/job/Container_Precheck_Velero/98/

Bug Number: DPCP-144

Signed-off-by: Swati Gupta <swgupta@swgupta-a01.vmware.com>