Changes to reflect update status in guest cluster snapshot CR.
The snapshot ID in the guest cluster used to contain the corresponding
supervisor cluster snapshot CR name. This was used to watch for
supervisor snapshot CRs to get the upload status in the guest. This
was recently changed and the snapshot ID no longer points to the
supervisor cluser snapshot CR.

Added a new field in Snapshot-Status that has the name of the supervisor
snapshot CR in the guest. We use the guest parameters map passed to the
paravirt snapshot function to return the supervisor snapshot CR name.
The rest of the code to watch for update status remains the same.

Signed-off-by: Swati Gupta <swgupta@swgupta-a01.vmware.com>