Changes to create a new namespace for backup-driver in the guest.
The velero-app-operator in the supervisor cluster writes the permissions to
access the supervisor namespace as a secret in a fixed namespace in the guest
cluster. Since the user can deploy velero in any namespace, we create a fixed
namespace in the guest cluster where the secret gets written. When the backup
driver container in the guest is started, it creates this fixed namespace and
waits for the seccret to become available.

Precheckin test: https://container-dp.svc.eng.vmware.com/job/Container_Precheck_Velero/401/

Signed-off-by: Swati Gupta <swgupta@swgupta-a01.vmware.com>