Check for the correct local mode feature flag when starting data manager pod. (#225)
On the vanilla cluster, the data manager install code checks for the feature
flag to determine if data manager pods need to be initialized. We were checking
for the wrong flag (local-mode) instead of the new EnableLocalMode flag. This
change fixes this issue.

Testing: Manual test with local mode on vanilla setup.

Signed-off-by: Swati Gupta <swgupta@vmware.com>