Move Snapshot/DeleteSnapshot/CloneFromSnapshot into Failed state on errors (#351)
In the current code, any failures that occur prior to the actual snapshot/clone/deletesnapshot operation do not move the associated CRDs into a Failed state. This leads to a situation where the plugin waits for either Completed or Failed state, but since the errors do not move the CRD into failed state, there is a possibility of the backup delete/create/restore operation becomes hung. The primary motive of this change is to move the CRDs into a terminal state.

Signed-off-by: Deepak Kinni <dkinni@vmware.com>